feat: Major system documentation update and comprehensive testing enhancement
- Updated .serena memory files with verified dependency relationships, error handling patterns, key entry points, MCP integration details, project overview, structure mapping, and testing patterns
- Enhanced benchmark suite with bottleneck profiler, comprehensive benchmarks, individual method benchmarks, and parallel execution benchmarks
- Added comprehensive test coverage including BM25 integration tests, compile time feature tests, embedding performance benchmarks, production verification tests, and accuracy testing
- Implemented unified project scope testing and symbol verification utilities
- Updated core search modules (BM25, fusion, symbol enhanced searcher, tantivy search, unified) with improved algorithms and error handling
- Enhanced storage implementations in LanceDB and simple vector database with better error management
- Improved embedding integration with Nomic model enhancements
- Added extensive documentation suite covering final system status, verification assessments, implementation plans, and technical analysis
- Included validation scripts and integration testing utilities
- Removed obsolete test database files and cleaned up project structure
- Added mission completion truth report documenting system state

This represents a comprehensive system analysis, documentation update, and testing infrastructure enhancement following SPARC methodology principles.

+++ src/search/fusion.rs
@@ -111,7 +111,7 @@ impl SimpleFusion {
                         None => {
                             return Err(SearchError::MissingSimilarityScore {
                                 file_path: semantic.file_path,
-                                chunk_index: semantic.chunk_index,
+                                chunk_index: semantic.chunk_index as u32,
                             });
                         }
                     };
@@ -212,7 +212,7 @@ impl SimpleFusion {
                         None => {
                             return Err(SearchError::MissingSimilarityScore {
                                 file_path: semantic.file_path,
-                                chunk_index: semantic.chunk_index,
+                                chunk_index: semantic.chunk_index as u32,
                             });
                         }
                     };
@@ -347,7 +347,7 @@ impl SimpleFusion {
                         None => {
                             return Err(SearchError::MissingSimilarityScore {
                                 file_path: semantic.file_path,
-                                chunk_index: semantic.chunk_index,
+                                chunk_index: semantic.chunk_index as u32,
                             });
                         }
                     };
